new obstacles
added tank terrain obstacles
made rockets produce lights
started on discharge orb

+++ Assets/Scripts/ECS/LightDataComponent.cs
@@ -8,5 +8,5 @@ public struct LightDataComponent : IComponentData
 {
     public float3 color;
     public float radius;
-    public int intensity;
+    public float intensity;
 }
